Rename function to more accurately describe function
